cypress/e2e/Regression/ServerSide/OnLoadTests/JSOnLoad2_Spec.ts fix
