+++ app/client/cypress/e2e/Regression/ServerSide/OnLoadTests/JSOnLoad2_Spec.ts
@@ -184,52 +184,6 @@ describe(
       agHelper.ValidateToastMessage("ran successfully", 0, 5);
     });
 
-    it("8. Tc 51, 52 Verify that JS editor function has a settings button available for functions marked async", () => {
-      jsEditor.CreateJSObject(
-        `export default {
-        myVar1: [],
-        myVar2: {},
-        myFun1: () => {	},
-        myFun2: async () => {	},
-        myFun3: async () => {	},
-        myFun4: async () => {	},
-        myFun6: async () => {	},
-        myFun7: async () => {	},
-        myFun8: () => {	},
-      }`,
-        {
-          paste: true,
-          completeReplace: true,
-          toRun: false,
-          shouldCreateNewJSObj: true,
-        },
-      );
-
-      jsEditor.VerifyAsyncFuncSettings("myFun2", false, false);
-      jsEditor.VerifyAsyncFuncSettings("myFun3", false, false);
-      jsEditor.VerifyAsyncFuncSettings("myFun4", false, false);
-      jsEditor.VerifyAsyncFuncSettings("myFun6", false, false);
-      jsEditor.VerifyAsyncFuncSettings("myFun7", false, false);
-
-      VerifyFunctionDropdown(
-        ["myFun1", "myFun8"],
-        ["myFun2", "myFun3", "myFun4", "myFun6", "myFun7"],
-      );
-
-      cy.get("@jsObjName").then((jsObjName) => {
-        jsName = jsObjName;
-        EditorNavigation.SelectEntityByName(
-          jsName as string,
-          EntityType.JSObject,
-        );
-        entityExplorer.ActionContextMenuByEntityName({
-          entityNameinLeftSidebar: jsName as string,
-          action: "Delete",
-          entityType: entityItems.JSObject,
-        });
-      });
-    });
-
     function AssertJSOnPageLoad(
       jsMethod: string,
       shouldCheckImport = false,
@@ -253,23 +207,6 @@ describe(
       deployMode.NavigateBacktoEditor();
       jsEditor.ConfirmationClick("No");
       agHelper.Sleep(2000);
-    }
-
-    function VerifyFunctionDropdown(
-      syncFunctions: string[],
-      asyncFunctions: string[],
-    ) {
-      cy.get(jsEditor._funcDropdown).click();
-      cy.get(jsEditor._funcDropdownOptions).then(function ($ele) {
-        expect($ele.eq(0).text()).to.be.oneOf(syncFunctions);
-        expect($ele.eq(1).text()).to.be.oneOf(asyncFunctions);
-        expect($ele.eq(2).text()).to.be.oneOf(asyncFunctions);
-        expect($ele.eq(3).text()).to.be.oneOf(asyncFunctions);
-        expect($ele.eq(4).text()).to.be.oneOf(asyncFunctions);
-        expect($ele.eq(5).text()).to.be.oneOf(asyncFunctions);
-        expect($ele.eq(6).text()).to.be.oneOf(syncFunctions);
-      });
-      cy.get(jsEditor._funcDropdown).click();
     }
   },
 );
