+++ app/client/cypress/limited-tests.txt
@@ -1,6 +1,6 @@
 # To run only limited tests - give the spec names in below format:
-cypress/e2e/Regression/ClientSide/Templates/Fork_Template_spec.js
-
+#cypress/e2e/Regression/ClientSide/Templates/Fork_Template_spec.js
+cypress/e2e/Regression/ServerSide/OnLoadTests/JSOnLoad2_Spec.ts
 
 # For running all specs - uncomment below:
 #cypress/e2e/**/**/*
